Move stellar-ledger specific things to ledger_impl
