+++ cmd/soroban-cli/src/signer/ledger.rs
@@ -1,12 +1,22 @@
 pub use ledger_impl::*;
 use crate::xdr::{
-    DecoratedSignature, Transaction,
+    self, DecoratedSignature, Transaction,
 };
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
     #[error("Ledger Device keys are not allowed: additional-libs feature must be enabled")]
     FeatureNotEnabled,
+
+    #[cfg(feature = "additional-libs")]
+    #[error(transparent)]
+    StellarLedger(#[from] stellar_ledger::Error),
+
+    #[error(transparent)]
+    TryFromSlice(#[from] std::array::TryFromSliceError),
+
+    #[error(transparent)]
+    Xdr(#[from] xdr::Error),
 }
 
 #[cfg(not(feature = "additional-libs"))]
@@ -55,6 +65,53 @@ mod ledger_impl {
 
 #[cfg(feature = "additional-libs")]
 mod ledger_impl {
+    use sha2::{Digest, Sha256};
+    use stellar_ledger::{Blob as _, Exchange, LedgerSigner};
+    use crate::xdr::{ Hash, Signature, SignatureHint };
+    use super::*;
+ 
+pub struct Ledger<T: Exchange> {
+    pub(crate) index: u32,
+    pub(crate) signer: LedgerSigner<T>,
+}
+
+impl<T: Exchange> Ledger<T> {
+    pub async fn sign_transaction_hash(
+        &self,
+        tx_hash: &[u8; 32],
+    ) -> Result<DecoratedSignature, Error> {
+        let key = self.public_key().await?;
+        let hint = SignatureHint(key.0[28..].try_into()?);
+        let signature = Signature(
+            self.signer
+                .sign_transaction_hash(self.index, tx_hash)
+                .await?
+                .try_into()?,
+        );
+        Ok(DecoratedSignature { hint, signature })
+    }
+
+    pub async fn sign_transaction(
+        &self,
+        tx: Transaction,
+        network_passphrase: &str,
+    ) -> Result<DecoratedSignature, Error> {
+        let network_id = Hash(Sha256::digest(network_passphrase).into());
+        let signature = self
+            .signer
+            .sign_transaction(self.index, tx, network_id)
+            .await?;
+        let key = self.public_key().await?;
+        let hint = SignatureHint(key.0[28..].try_into()?);
+        let signature = Signature(signature.try_into()?);
+        Ok(DecoratedSignature { hint, signature })
+    }
+
+    pub async fn public_key(&self) -> Result<stellar_strkey::ed25519::PublicKey, Error> {
+        Ok(self.signer.get_public_key(&self.index.into()).await?)
+    }
+}
+
     #[cfg(all(feature = "additional-libs", not(feature = "emulator-tests")))]
     pub async fn new(hd_path: u32) -> Result<Ledger<stellar_ledger::TransportNativeHID>, Error> {
         let signer = stellar_ledger::native()?;
